feat: move determine_user_role function to accounts/utils.py
