+++ accounts/utils.py
@@ -8,6 +8,7 @@ from unidecode import unidecode
 from accounts.models import UserAffiliation
 from communities.models import Community
 from institutions.models import Institution
+from projects.models import ProjectCreator
 from researchers.models import Researcher
 
 
@@ -108,3 +109,25 @@ def remove_user_from_affiliated_communities_and_institutions(user: User, affilia
 
     for institution in affiliation.institutions.all():
         remove_user_from_account(user, institution)
+
+
+def determine_user_role(user: User) -> str:
+    is_account_creator = Researcher.objects.filter(user=user).exists() or \
+                             Community.objects.filter(community_creator=user).exists() or \
+                             Institution.objects.filter(institution_creator=user).exists()
+
+    is_project_creator = ProjectCreator.objects.filter(project__project_creator__id=user.id).exists()
+    if is_account_creator or is_project_creator:
+        return 'is_creator_or_project_creator'
+
+    is_member = Community.objects.filter(admins__id__contains=user.id).exists() or \
+                            Community.objects.filter(editors__id__contains=user.id).exists() or \
+                            Community.objects.filter(viewers__id__contains=user.id).exists() or \
+                            Institution.objects.filter(admins__id__contains=user.id).exists() or \
+                            Institution.objects.filter(editors__id__contains=user.id).exists() or \
+                            Institution.objects.filter(viewers__id__contains=user.id).exists()
+
+    if is_member > 0:
+        return 'is_member'
+
+    return 'default'
